Added zoom to fit locations behavior
