+++ src/Map.jsx
@@ -1,51 +1,16 @@
-import React, { useEffect, useLayoutEffect, useRef, useState } from "react";
-import { MapContainer, TileLayer, GeoJSON, useMap } from "react-leaflet";
-import { Box, Button } from "grommet";
-import { Add, Grommet, Subtract, Waypoint } from "grommet-icons";
+import React, { useEffect, useRef, useState } from "react";
+import { MapContainer, TileLayer, GeoJSON } from "react-leaflet";
+import { Box } from "grommet";
+import { Grommet } from "grommet-icons";
 
 import GrommetMarker from "./GrommetMarker";
 import MarkerClusterGroup from "./MarkerClusterGroup";
 import { generateLocations } from "./utils/locations";
 import { geojsonLocations } from "./utils/geojson_data";
+import { Controls } from "./Controls";
 
 const initialStyle = {
   height: "100%",
-};
-
-const ZoomControl = ({ home }) => {
-  const map = useMap();
-  return (
-    <div className="leaflet-bottom leaflet-right">
-      <div className="leaflet-control">
-        <Box elevation="large" round="medium" background="background">
-          <Button
-            icon={<Add />}
-            onClick={() => {
-              map.zoomIn();
-            }}
-          />
-          <Button
-            icon={<Subtract />}
-            onClick={() => {
-              map.zoomOut();
-            }}
-          />
-          <Box
-            border={{
-              color: "border-weak",
-              side: "top",
-            }}
-          />
-          <Button
-            icon={<Waypoint />}
-            onClick={() => {
-              map.panTo(home);
-            }}
-          />
-        </Box>
-      </div>
-    </div>
-  );
 };
 
 function Map() {
@@ -90,7 +55,7 @@ function Map() {
               &copy; <a href="http://openstreetmap.org">OpenStreetMap</a> contributors`}
             url="https://tiles.stadiamaps.com/tiles/alidade_smooth/{z}/{x}/{y}{r}.png"
           />
-          <ZoomControl home={geolocation} />
+          <Controls home={geolocation} locations={geojsonLocations} />
           <GrommetMarker position={geolocation} icon={<Grommet />} />
           <MarkerClusterGroup>
             {generateLocations(50, { center: geolocation, radius: 5 }).map(
